FP refactor of non-sim action dispatch
