+++ src/index.js
@@ -16,7 +16,6 @@ import {
     UPDATE_FREQ_SIM,
 } from './const_vals.js';
 
-import { doNonSimUpdate } from './do_nonsim_update.js';
 import { applyArgChain } from './util.js';
 
 import {
@@ -31,6 +30,8 @@ import {
     mutableRender
 } from './reduxlike/action_creators.js';
 
+import { actionGroup_NonsimActions } from './reduxlike/actiongroup_nonsimactions.js';
+
 import {
     simGetSavedClock,
     simGetRunning,
@@ -40,26 +41,31 @@ import {
 import { storeInit } from './reduxlike/store_init.js';
 
 
-// *** Define function-chaining function applied to our store action dispatcher
+// *** Define argument-chaining function applied to our store action dispatcher
 const applyArgChainActionDispatch = applyArgChain(actionDispatch);
 
 
 // ***********************************************************************************
 // *** Code that actually does stuff
 
-// create a reference to an initialized store object
+// create our store by referencing an initialized store object
 export let myStore = storeInit(
     document.getElementById(CREATURE_TIME_CHART).getContext('2d'),
     document.getElementById(CREATURE_GEO_CHART).getContext('2d'),
     document.getElementById(CREATURE_STATUS_BOX)
 );
 
-// change the sim status to running
-myStore = actionDispatch(myStore)(startSim());
+// dispatch a series of actions to our store
+myStore = applyArgChainActionDispatch(myStore)(
+    // change the sim status to running
+    startSim(),
 
-// do the initial non-sim draw
-myStore = doNonSimUpdate(myStore);
-myStore = actionDispatch(myStore)(mutableRender());
+    // dispatch non-sim-related actions such as queuing initial chart draws
+    actionGroup_NonsimActions(myStore),
+
+    // do the initial UI draws
+    mutableRender()
+);
 
 // start repeatedly updating our application at sim frequency
 let requestId = setInterval(appUpdate, UPDATE_FREQ_SIM);
@@ -76,30 +82,46 @@ function appUpdate() {
         simGetRunning(myStore) &&
         (!storeIsLocked(myStore))
     ) {
-        // yes: set store lock, do physType act, advance sim, unset store lock
+        // yes: dispatch a series of actions to the store to update the sim
         myStore = applyArgChainActionDispatch(myStore)(
+            // set store lock
             lockStore(),
+
+            // do physType acts
             myStore.physTypeStore.map(
                 (this_physType, i) => doPhysTypeAct(this_physType)(i)
             ),
+
+            // advance sim
             advanceSim(),
+
+            // unset store lock
             unlockStore()
         );
     }
 
     // has UPDATE_FREQ_NONSIM time passed since last non-sim update
-    //  AND store lock not set?
+    //  and store lock not set?
     if (
         (performance.now() > (simGetSavedClock(myStore) + UPDATE_FREQ_NONSIM)) &&
         (!storeIsLocked(myStore))
     ) {
-        // update the non-sim parts of our app store
-        myStore = doNonSimUpdate(myStore);
+        // yes: dispatch a series of actions to the store to update the non-sim stuff
+        myStore = applyArgChainActionDispatch(myStore)(
+            // set store lock
+            lockStore(),
 
-        // render the application
-        myStore = actionDispatch(myStore)(mutableRender());
+            // update the non-sim parts of our app store
+            actionGroup_NonsimActions(myStore),
 
-        // remember the current time
-        myStore = actionDispatch(myStore)(saveClockForSim(performance.now()));
+            // render the application
+            mutableRender(),
+
+            // remember the current time
+            saveClockForSim(performance.now()),
+
+            // unset store lock
+            unlockStore()
+        )
     }
 };
